requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
httpx==0.25.1
beautifulsoup4==4.12.2
dateparser==1.2.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
tinydb==4.8.0
